[Mac] move automated test projects into new Common/Mac/TestProjects folder
