--- tests/common/mac/BasicPCLTest/MyLibrary/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Newtonsoft.Json" version="6.0.1" targetFramework="net45" />
-  <package id="NServiceBus" version="5.1.0" targetFramework="net45" />
-  <package id="NServiceBus.Newtonsoft.Json" version="0.1.0" targetFramework="net45" />
-</packages>
