+++ tests/common/mac/TestProjects/BasicPCLTest/BasicPCLTest/AppDelegate.cs
