+++ tests/common/mac/TestProjects/BasicPCLTest/BasicPCLTest/BasicPCLTest.csproj
@@ -61,11 +61,11 @@
   <ItemGroup>
     <None Include="Info.plist" />
   </ItemGroup>
-  <Import Project="$(MSBuildExtensionsPath)\Xamarin\Mac\Xamarin.Mac.CSharp.targets" />
   <ItemGroup>
     <ProjectReference Include="..\MyLibrary\MyLibrary.csproj">
       <Project>{61393B2B-B0AF-4875-84E4-742DCCB87749}</Project>
       <Name>MyLibrary</Name>
     </ProjectReference>
   </ItemGroup>
+  <Import Project="$(MSBuildExtensionsPath)\Xamarin\Mac\Xamarin.Mac.CSharp.targets" />
 </Project>
