+++ tests/common/mac/TestProjects/BasicPCLTest/BasicPCLTest/Main.cs
@@ -2,13 +2,14 @@
 
 using AppKit;
 
-namespace ServiceModel_Test
+namespace BasicPCLTest
 {
 	static class MainClass
 	{
 		static void Main (string[] args)
 		{
 			NSApplication.Init ();
+			NSApplication.SharedApplication.Delegate = new AppDelegate ();
 			NSApplication.Main (args);
 		}
 	}
