+++ tests/common/mac/TestProjects/BasicPCLTest/MyLibrary/MyClass.cs
