+++ tests/common/mac/TestProjects/BasicPCLTest/MyLibrary/Properties/AssemblyInfo.cs
