+++ tests/common/mac/TestProjects/ImmutableCollection_Test/ImmutableCollection_Test/Entitlements.plist
