+++ tests/common/mac/TestProjects/ImmutableCollection_Test/ImmutableCollection_Test/ImmutableCollection_Test.csproj
@@ -65,8 +65,5 @@
     <Compile Include="Main.cs" />
     <Compile Include="AppDelegate.cs" />
   </ItemGroup>
-  <ItemGroup>
-    <InterfaceDefinition Include="Main.storyboard" />
-  </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Mac\Xamarin.Mac.CSharp.targets" />
 </Project>
