+++ tests/common/mac/TestProjects/ImmutableCollection_Test/ImmutableCollection_Test/Main.cs
@@ -5,6 +5,7 @@ namespace ImmutableCollection_Test {
 		static void Main (string [] args)
 		{
 			NSApplication.Init ();
+			NSApplication.SharedApplication.Delegate = new AppDelegate ();
 			NSApplication.Main (args);
 		}
 	}
