+++ tests/common/mac/TestProjects/Protobuf_Test/Protobuf_Test/AppDelegate.cs
