+++ tests/common/mac/TestProjects/Protobuf_Test/Protobuf_Test/Info.plist
