+++ tests/common/mac/TestProjects/Protobuf_Test/Protobuf_Test/Main.cs
@@ -2,13 +2,14 @@
 
 using AppKit;
 
-namespace BasicPCLTest
+namespace Protobuf_Test
 {
 	static class MainClass
 	{
 		static void Main (string[] args)
 		{
 			NSApplication.Init ();
+			NSApplication.SharedApplication.Delegate = new AppDelegate ();
 			NSApplication.Main (args);
 		}
 	}
