+++ tests/common/mac/TestProjects/Protobuf_Test/Protobuf_Test/Protobuf_Test.csproj
