+++ tests/common/mac/TestProjects/ServiceModel_Test/ServiceModel_Test/AppDelegate.cs
