+++ tests/common/mac/TestProjects/ServiceModel_Test/ServiceModel_Test/Main.cs
@@ -2,13 +2,14 @@
 
 using AppKit;
 
-namespace Protobuf_Test
+namespace ServiceModel_Test
 {
 	static class MainClass
 	{
 		static void Main (string[] args)
 		{
 			NSApplication.Init ();
+			NSApplication.SharedApplication.Delegate = new AppDelegate ();
 			NSApplication.Main (args);
 		}
 	}
